oasis/requirements.txt
grpcio==1.73.1
grpcio-tools==1.62.3
jupyter==1.1.1
icos-fl @ git+https://github.com/anaskalt/icos-fl.git
pandas>=1.5.3
flwr[simulation]>=1.18
torch==2.5.1
wandb==0.19.8
scikit-learn>=1.6.1
mlflow==2.22.0
seaborn== 0.13.2
matplotlib==3.7.1
tensorflow-cpu==2.15.0
pydantic==2.11.4
matplotlib==3.7.1
tensorflow-cpu==2.15.0
scipy==1.10.1
statsmodels==0.14.0
scikit-learn>=1.6.1
xgboost==1.7.6
tqdm==4.65.0
category_encoders==2.6.3
shap==0.46.0
transformers==4.30.0
numpy>=1.24.4
protobuf==4.24.4
bentoml==1.1.11
nannyml==0.13.0
monotonic==1.6